+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Release History
 - Add docs for `wsproto.Connection`
 - Add support for Python 3.12, 3.13, and 3.14.
 - Drop support for Python 3.7, 3.8, and 3.9.
+- Improve Python typing, specifically bytes vs. bytearray.
 - Various linting, styling, and packaging improvements.
 
 1.2.0 (2022-08-23)